requirements.txt
m3u8==0.5.3
requests==2.22.0
pycryptodome==3.9.6
